test: merge language switching tests with basic usage
