+++ specs/fixtures/basic_usage/app.vue
@@ -1,9 +1,35 @@
 <script setup lang="ts">
-const head = useLocaleHead({ addSeoAttributes: true })
+import { useI18n } from '#i18n'
+import { useRuntimeConfig } from 'nuxt/app'
+
+const { finalizePendingLocaleChange } = useI18n()
+
+const skipSettingLocale = useRuntimeConfig().public.i18n.skipSettingLocaleOnNavigate
+const pageTransition = {
+  name: 'my',
+  mode: 'out-in',
+  onBeforeEnter: async () => {
+    await finalizePendingLocaleChange()
+  }
+}
 </script>
 
 <template>
   <NuxtLayout>
-    <NuxtPage />
+    <NuxtPage id="nuxt-page" :transition="skipSettingLocale ? pageTransition : undefined" />
   </NuxtLayout>
 </template>
+
+<style>
+section {
+  margin: 1rem 0;
+}
+.my-enter-active,
+.my-leave-active {
+  transition: opacity 0.3s;
+}
+.my-enter,
+.my-leave-active {
+  opacity: 0;
+}
+</style>
