+++ specs/fixtures/basic_usage/pages/greetings.vue
