+++ specs/fixtures/basic_usage/pages/post/[id].vue
