+++ specs/fixtures/basic_usage/pages/user/profile.vue
@@ -1,3 +1,5 @@
 <template>
-  <p id="profile-page">This is profile page</p>
+  <div>
+    <p id="profile-page">This is profile page</p>
+  </div>
 </template>
