--- specs/fixtures/switcher/i18n.config.ts
@@ -1,18 +1,0 @@
-export default {
-  legacy: false,
-  messages: {
-    fr: {
-      home: 'Accueil',
-      about: 'À propos',
-      posts: 'Articles',
-      dynamic: 'Dynamique'
-    },
-    en: {
-      home: 'Homepage',
-      about: 'About us',
-      posts: 'Posts',
-      dynamic: 'Dynamic'
-    }
-  },
-  fallbackLocale: 'en'
-}
